split directional shadow renderer and introduce shadow maps aggregate
